Style
Did a little more styling, got everything more how I want it to be.
Project Title at the top, a menu-bar of tools [most of which were just there for me as a developer and will probably get cleared away and replaced], and the planner at the top to fabricate account objects.

Added Net Worth stat and a balance checking function to back it up. Rendering the account table now updates "Net worth", though I will probably shift NW to being a portfolio property that is read/set by the function to make way for saving, loading, and exporting portfolios.