Saving
Moved global portfolio array to property of global object f.

Added a fair amount of infrastructure to support encoding and decoding account objects from JSON, or in other words, making it possible to save app state to local storage, giving people the ability to load, save, and reset portfolio profiles.

Portfolios are now autosaved during every re-rendering of the account-table.

Added Lodash to dependencies, but it's not being used for anything right now..