Rendering Overhaul
I went through with my idea for card selection. Now instead of rendering a table individual accounts are looked at on a card-by-card basis. I think I will revisit the card table (And other financial views) when I've developed more of the underlying infrastructure I want.

-Accounts now have removability. This is a portfolio object method that takes one argument - an array index.

- Assets now have removability, both by name (Though this is not in use), and by clever html trickiness.

-Debt objects were formalized and given a debt Factory and debt constructor object. Right now they just call a new instance of a bank account, but with a negative balance & APR.

Render refactored to take an argument, which is the global variable "x". X is the currently selected card, ergo, it is an array index. This is probably bad design and I'll probably make it a property of my global object f to prevent namespace issues.

- Account Selector moved and slightly styled. I experimented with it in a few places and decided I liked it up in the info panel. Interested in finding a more user friendly way to design it, but for now, I'm satisfied.